requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.1
tqdm==4.66.1
lxml==4.9.3
urllib3==2.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6 